[Winlogbeat] Move winlogbeat javascript processor to libbeat (#21402)
* Move winlogbeat processor to libbeat

- needed for filebeat winlog input
- changed name to windows
- also register as winlogbeat for backwards compatibility

(cherry picked from commit fd52ca5adff8a68370e98ae89e101e4060776020)
