+++ libbeat/processors/script/javascript/module/windows/windows_test.go
@@ -17,7 +17,7 @@
 
 // +build windows
 
-package winlogbeat
+package windows
 
 import (
 	"testing"
